api/requirements.txt
werkzeug==2.3.8
flask==2.3.3
flask-restful==0.3.9
flask-login==0.6.2
flask-mysql==1.5.2
cryptography==38.0.1
python-dotenv==1.0.1
numpy==1.26.4
flask-cors==4.0.0
